Change method return type and add increment/getter

+++ code/src/main/java/com/onlinecontacttracing/storage/data-storage/PositiveUser.java
@@ -42,7 +42,7 @@ public class PositiveUser {
     return userEmail;
   }
   
-  public long getNumberOfEmailsSent() {
+  public int getNumberOfEmailsSent() {
     return emailsSent;
   }
